Add snapshot tests, refactor button
Previously, the `<Button>` component took an `icon` prop that would itself be a component. This was a little awkward. In addition, the loader is technically not an icon. So, we'd need a set an additional property for that.

But, they are both just children of the button. So, I opted to treat them as such.

We'll need to update the storybook as well, but I wanted to get this change in front of some eyes first.