Add support for the data-tooltip-length attribute. #patch (#111)
* Add support for the data-tooltip-length attribute. #110

See issue #110.

* Support data-tooltip-length on HTMLTooltips

* Add missing attribute in story

* Remove support for 'fit'

The "fit" style is broken in styleguide currently. Not sure we even need that option, since the default tooltip expands to fit available content.

See: https://github.com/sendgrid/style-guide/issues/165

* Improve content in stories

* Use proper variable.

* Add anti-story for style bleedthrough.

Tooltips often (but shouldn't) inherit styles of the parent. These styles tend to be typographic.

* Add newline.

* Handle case when the attr `direction` is not supplied.

It should use "up" as the default. Note: we should probably handle the absence of this attribute in CSS.

* Add better title.

* Add position.

Allows this to work everywhere without being wrapped outside of the component.

* Consolidate standard tooltip stories and include all permutations.

* Consolidate bleed story into all standard stories.

* Consolidated HTML tooltip stories and added permutations.

* Reformat presentation.

* Added permutations to underline story.

* update package-lock.json

* Update changelog.

Adding support for this attribute seems to fit the criteria of a backwards-compatible functional improvement.