MCFE-479 First pass as responsive dropdown menu #patch (#204)
* First pass as responsive dropdown menu

This is very much an early idea that I'd love to get some feedback on.

**The Problem**: Menus get cut off when there is not enough room left in the window for them.

**Solution**: Open menus in the other direction when there is not enough room.

There are a few caveats here. First, I'm just using the minimum width that we *know* a menu is going to take up. It's obvious hard to know the size of your children in either React or CSS. I don't know of many menus that blow past the minimum. So, my gut says let's go with this approach for now before we over optimize. (For example, we could wait for the children to mount and then tell us how big they are, but this is non-trivial.)

`getBoundingClientRect()` isn't the cheapest function in the world, but we're also not doing anything crazy like calling it on every frame. It's only in response to a user action. So, I think the performance impacts will be infinitesimally small.

* Update snapshots